feat(map): geolocation usecase, entity 분리
