+++ packages/entity/src/geolocation.ts
@@ -1,0 +1,7 @@
+import type { MapPosition } from './map';
+
+export interface GeolocationController {
+  getCurrentPosition(): Promise<MapPosition>;
+  startWatching(options?: PositionOptions): Promise<MapPosition>;
+  stopWatching(): void;
+}
