+++ packages/infrastructures/src/controllers/geolocationController.ts
@@ -9,19 +9,41 @@ export default class GeolocationController {
     this.filters = filters;
   }
 
-  getInitialPosition(): Promise<MapPosition> {
+  private applyFilters(
+    position: MapPosition,
+    accuracy: number,
+  ): Promise<MapPosition> | MapPosition {
+    return this.filters.reduce<Promise<MapPosition> | MapPosition>(
+      async (pos, filter) => filter.filter(await pos, accuracy),
+      position,
+    );
+  }
+
+  getCurrentPosition(): Promise<MapPosition> {
     return new Promise((resolve, reject) => {
       if (navigator.geolocation) {
         navigator.geolocation.getCurrentPosition(
           (pos) => {
             if (process.env.NEXT_PUBLIC_USE_API_MOCKING === 'true') {
-              const latitude = 37.498095;
-              const longitude = 127.028979;
-              resolve({ latitude, longitude });
+              // const latitude = 37.498095;
+              // const longitude = 127.028979;
+              const latitude = pos.coords.latitude;
+              const longitude = pos.coords.longitude;
+
+              const position = this.applyFilters(
+                { latitude, longitude },
+                pos.coords.accuracy,
+              );
+
+              resolve(position);
             } else {
               const latitude = pos.coords.latitude;
               const longitude = pos.coords.longitude;
-              resolve({ latitude, longitude });
+              const position = this.applyFilters(
+                { latitude, longitude },
+                pos.coords.accuracy,
+              );
+              resolve(position);
             }
           },
           (err) => {
@@ -34,9 +56,9 @@ export default class GeolocationController {
             reject(err);
           },
           {
-            enableHighAccuracy: false,
-            maximumAge: 180000,
-            timeout: 7000,
+            enableHighAccuracy: true,
+            timeout: 15000,
+            maximumAge: 0,
           },
         );
       } else {
@@ -45,66 +67,40 @@ export default class GeolocationController {
     });
   }
 
-  private applyFilters(position: MapPosition, accuracy: number): MapPosition {
-    return this.filters.reduce(
-      (pos, filter) => filter.filter(pos, accuracy),
-      position,
-    );
-  }
-
-  getCurrentPosition(): Promise<MapPosition> {
+  startWatching(options?: PositionOptions): Promise<MapPosition> {
     return new Promise((resolve, reject) => {
-      if (!navigator.geolocation) {
-        reject(new Error('notSupport'));
-        return;
+      if (navigator.geolocation) {
+        this.watchId = navigator.geolocation.watchPosition(
+          (pos) => {
+            const position = {
+              latitude: pos.coords.latitude,
+              longitude: pos.coords.longitude,
+            };
+            resolve(this.applyFilters(position, pos.coords.accuracy));
+          },
+          (err) => {
+            if (err.code === 3) {
+              reject(new Error('delayed'));
+            }
+            if (err.code === 1) {
+              reject(new Error('blocked'));
+            }
+            reject(err);
+          },
+          {
+            ...options,
+            enableHighAccuracy: true, // 높은 정확도
+            timeout: 15000, // 위치를 가져오는 제한 시간
+            maximumAge: 0, // 캐시된 위치 정보를 사용하지 않음
+          },
+        );
       }
-
-      navigator.geolocation.getCurrentPosition(
-        (pos) => {
-          const position = {
-            latitude: pos.coords.latitude,
-            longitude: pos.coords.longitude,
-          };
-          resolve(this.applyFilters(position, pos.coords.accuracy));
-        },
-        (err) => reject(new Error(err.message)),
-        {
-          enableHighAccuracy: true,
-          timeout: 15000,
-          maximumAge: 0,
-        },
-      );
     });
-  }
-
-  startWatching(
-    onSuccess: (position: MapPosition) => void,
-    onError: (error: GeolocationPositionError) => void,
-    options?: PositionOptions,
-  ): void {
-    if (navigator.geolocation) {
-      this.watchId = navigator.geolocation.watchPosition(
-        (pos) => {
-          const position = {
-            latitude: pos.coords.latitude,
-            longitude: pos.coords.longitude,
-          };
-          onSuccess(this.applyFilters(position, pos.coords.accuracy));
-        },
-        onError,
-        {
-          enableHighAccuracy: true,
-          timeout: 15000,
-          maximumAge: 0,
-          ...options,
-        },
-      );
-    }
   }
 
   stopWatching(): void {
     if (this.watchId !== null) {
-      navigator.geolocation.clearWatch(this.watchId);
+      navigator.geolocation.clearWatch(this.watchId); // 위치 추적을 중지
       this.watchId = null;
     }
   }
